Fix dtype mismatch in projection, add V15 diagnostic results
W stays float32 on GPU (precision matters for projection).
latent_to_soft_prompt handles device+dtype matching efficiently.

V15 diagnostic: all conditions 95% at depth-2 (ceiling effect).
Tasks too easy for Qwen3-4B - needs harder tasks to differentiate.

+++ experiments/run_v15_geometry_isolation.py
@@ -87,7 +87,7 @@ def main():
     num_soft_tokens = 8
     d_out = num_soft_tokens * embed_dim
     W = make_row_orthonormal_W(d_latent, d_out, seed=1234)
-    W = W.to(device=encoder._device, dtype=encoder.model.dtype)
+    W = W.to(device=encoder._device)  # Keep float32 for precision
     print(f"W shape: {W.shape} (d_latent={d_latent}, d_out={d_out})", flush=True)
     print(f"W device: {W.device}", flush=True)
